feat: split file
move installDependencies

add getBlockListFromGit
